[style] likePreview 관련 함수 모으기
